backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
elasticsearch
minio
tensorflow
torch
spacy
opencv-python
pytesseract
Pillow
spacy
